+++ tests/testthat/test-test_indirect_lm.R
@@ -20,7 +20,7 @@ sim_only <- power4test(nrep = 5,
                        n = 100,
                        R = 50,
                        ci_type = "boot",
-                       fit_model_args = list(fit_function = lmhelprs::many_lm),
+                       fit_model_args = list(fit_function = "lm"),
                        do_the_test = FALSE,
                        iseed = 1234)