+++ tests/testthat/test-test_moderation_lm.R
@@ -18,7 +18,7 @@ sim_only <- power4test(nrep = 5,
                        model = mod,
                        pop_es = mod_es,
                        n = 100,
-                       fit_model_args = list(fit_function = lmhelprs::many_lm),
+                       fit_model_args = list(fit_function = "lm"),
                        do_the_test = FALSE,
                        iseed = 1234)